Profile==NULL assert fix

+++ Graphs/glcommon.h
@@ -22,7 +22,7 @@ extern QFont * bigfont;
 
 class gGraphWindow;
 void GetTextExtent(QString text, float & width, float & height, QFont *font=defaultfont);
-void DrawText(gGraphWindow & wid, QString text, float x, float y, float angle=0, QColor color=QColor("black"),QFont *font=defaultfont);
+void DrawText(gGraphWindow & wid, QString text, int x, int y, float angle=0, QColor color=QColor("black"),QFont *font=defaultfont);
 
 void LinedRoundedRectangle(int x,int y,int w,int h,int radius,int lw,QColor color);
 void RoundedRectangle(int x,int y,int w,int h,int radius,const QColor color);
